Fix documentation for PostgreSQL `DB_TYPE` value
This commit fixes the documentation for the PostgreSQL environment
variable values. Previously the documentation stated that the expected
`DB_TYPE` value when using PostgreSQL was "pgsql", but the `DBType` enum
actually expects a name of "postgres" which then maps to a value of
"pgsql".

It also fixes a small formatting mistake in the README.md file.